hackerrank plus minous program using java

+++ Precision.java
@@ -2,7 +2,7 @@ public class Precision {
     public static void main(String[] args) {
         double val = 9.0;
         System.out.println("Precision " + String.format("%.10f", val));
-        double m = 99;
-        System.out.println(String.format("%.10f", m)); //floating point "%.10f" 
+        float m = 99;
+        System.out.println(String.format("%.10f", m)); //floating point "%.10f"
     }
 }
